refactor(tests): promote fixtures for reuse
move the fixtures introduced in acde492f6d89de9f36c6807f0336d0bc2b78ee43
to `tests.command_line.conftest` so that they can be reused by other
test modules
